Homework 1 answers added.

+++ Session-1/homework/Ali-Izadi/6.py
@@ -1,1 +1,1 @@
-print("Result Of Correct Division Is: ", 12.54//3.15)
+print("Result Of Correct Division Is: ", (int)(12.54//3.15))
